requirements.txt
# Python dependencies
numpy
opencv-python
ultralytics
pybind11
pyyaml

# Note: The following must be installed separately (not via pip):
# - CUDA Toolkit (11.x or 12.x)
# - TensorRT (8.x or 10.x)
# - OpenCV C++ libraries (for building the TRT detector)

